делает 5 дз
