+++ js/setup.js
@@ -1,0 +1,31 @@
+const DESCRIPTION_PHOTO = [
+  'Лучшее фото',
+  'Самое лучшее фото',
+  'Просто фото',
+  'Не самое удачное фото',
+  'Самое неудачное фото',
+  'Удачное фото',
+  'Самое удачное фото',
+  'Лучшее фото',
+  'Самое лучшее фото',
+  'Просто фото',
+  'Не самое удачное фото',
+  'Самое неудачное фото',
+  'Удачное фото',
+  'Самое удачное фото',
+  'Лучшее фото',
+  'Самое лучшее фото',
+  'Просто фото',
+  'Не самое удачное фото',
+  'Самое неудачное фото',
+  'Удачное фото',
+  'Самое удачное фото',
+  'Самое неудачное фото',
+  'Удачное фото',
+  'Самое удачное фото',
+  'Просто фото',
+];
+
+const SIMILAR_PHOTO_COUNT = 25;
+
+export {DESCRIPTION_PHOTO, SIMILAR_PHOTO_COUNT};
